- Fixing the bug related to UID (reported by Barry). - Fixing the support for charting double values (previously it could  be used for producing graphs using INTEGER values only).

+++ src/gsn/vsensor/ChartVirtualSensor.java
@@ -255,7 +255,7 @@ class ChartInfo {
             dataCollectionForTheChart.addSeries( timeSeries );
          }
          try {
-            timeSeries.addOrUpdate( new FixedMillisecond( new Date( streamElement.getTimeStamp( ) ) ) , Integer.parseInt( streamElement.getData( )[ i ].toString( ) ) );
+            timeSeries.addOrUpdate( new FixedMillisecond( new Date( streamElement.getTimeStamp( ) ) ) , Double.parseDouble( streamElement.getData( )[ i ].toString( ) ) );
          } catch ( SeriesException e ) {
             logger.warn( e.getMessage( ) , e );
          }
